Cleared the old html2canvas constants (no longer used anywhere).
